Drop typelevel-prelude dependency
We aren't actually using anything provided by typelevel-prelude;
instead, we're only using its re-exports. Dropping the dependency and
importing RLProxy and the Union class from where they are defined means
that we have fewer dependencies to worry about, so version
incompatibilities are less likely to occur.

I'd suggest releasing this as a patch-level change; previously we have
recommended releasing changes in which dependencies have been dropped as
major-level, because it can break downstream libraries if they are
relying on the presence of that dependency, but I have come to believe
that this policy is worse since it results in far too many unnecessary
major version bumps, and downstream libraries depending on a particular
transitive dependency but not declaring that dependency are arguably
broken anyway.